+++ lib/folio/engine.rb
@@ -81,6 +81,9 @@ module Folio
       { site_id: controller.site_for_new_files.id }
     }
 
+    config.folio_console_links_mapping = {}
+    config.folio_console_links_additional_filters = {}
+
     config.folio_content_templates_editable = false
 
     config.folio_leads_from_component_class_name = nil